nicesrc: fix deadlock in create funcion
There is a deadlock under some race conditions caused for the
cond_wait into the "create" function.
The race condition is typically caused if the cond_signal
(in "stop" function) is run before the cond_wait in "create"
function and any data is received in the read_callback.

This patch fixes the problem implementing virtual GstBaseSrc
methods properly.

According to the GstBaseSrc doc:
unlock ()
  Unlock any pending access to the resource.
  Subclasses should unblock any blocked function ASAP.
  In particular, any create() function in progress should be
  unblocked and should return GST_FLOW_FLUSHING.
  Any future GstBaseSrcClass.create() function call should
  also return GST_FLOW_FLUSHING until the GstBaseSrcClass.unlock_stop()
  function has been called.

unlock_stop ()
  Clear the previous unlock request.
  Subclasses should clear any state they set during
  GstBaseSrcClass.unlock(), such as clearing command queues.

Change-Id: I4d4b2f1814b7c85b65bf2c2095fce52b77e785a4