debian/rules: Disable assert (#define NDEBUG 1)
Runs ./configure with '--disable-assert', which will #define the
standard macro NDEBUG; this tells the assert() macros that the process
should not be aborted whenever the checked expression is false.

This is desirable because libnice uses lots of assert() and we've
suffered several crashes already due to this. It's better to leave a
single connection in a bad state, than to throw away the whole process
and all of the sessions that had been successfully established.