PROCALL-895 memleak - free agent->refresh_list (79928 Bytes per every call)
based on Fix from caf9f1d8 19.12.2020

* agent: avoid leak of all turn refreshes when disposing the agent

With this patch we free all outstanding refreshes when the agent dispose
method is called, even those that are in the way to be discarded
asynchronously, when a stream is removed.

We also make the final user callback of the refresh proces synchronous,
so we don't have to deal with an heap use-after-free problem. This also
requires to order some parts of code.